+++ src/main/java/com/bae/persistence/repository/ActivityRepository.java
@@ -13,9 +13,11 @@ public interface ActivityRepository {
 	String getAllActivitiesByCategory(String userEmail, String category);
 
 	// Update
-	String updateActivity(String userEmail, String activityLog, Long id);
+
+	String updateActivity(String userEmail, String activityLog, int id);
 
 	// Delete
-	String deleteActivity(String userEmail, Long id);
+	String deleteActivity(String userEmail, int id);
+
 
 }
